Move consuming thread to WatcherRegistry
